// shadow story/src/types/performanceSkeleton.ts
export type PerformanceBoneMetrics = {
  skinScale: number;
  skinScaleHead: number;
  skinScaleTorso: number;
  skinScalePelvis: number;
  skinScaleLeftUpperArm: number;
  skinScaleLeftForearm: number;
  skinScaleRightUpperArm: number;
  skinScaleRightForearm: number;
  skinScaleLeftLeg: number;
  skinScaleRightLeg: number;
  offsetHeadX: number;
  offsetHeadY: number;
  offsetTorsoX: number;
  offsetTorsoY: number;
  offsetPelvisX: number;
  offsetPelvisY: number;
  offsetLeftUpperArmX: number;
  offsetLeftUpperArmY: number;
  offsetLeftForearmX: number;
  offsetLeftForearmY: number;
  offsetRightUpperArmX: number;
  offsetRightUpperArmY: number;
  offsetRightForearmX: number;
  offsetRightForearmY: number;
  offsetLeftLegX: number;
  offsetLeftLegY: number;
  offsetRightLegX: number;
  offsetRightLegY: number;
  jointShoulderX: number;
  jointShoulderY: number;
  jointElbowX: number;
  jointElbowY: number;
  jointHipX: number;
  jointHipY: number;
  jointKneeX: number;
  jointKneeY: number;
  jointLeftShoulderX: number;
  jointLeftShoulderY: number;
  jointRightShoulderX: number;
  jointRightShoulderY: number;
  jointLeftElbowX: number;
  jointLeftElbowY: number;
  jointRightElbowX: number;
  jointRightElbowY: number;
  jointLeftHipX: number;
  jointLeftHipY: number;
  jointRightHipX: number;
  jointRightHipY: number;
  jointLeftKneeX: number;
  jointLeftKneeY: number;
  jointRightKneeX: number;
  jointRightKneeY: number;
  jointNeckRodX: number;
  jointNeckRodY: number;
  jointRearHandRodX: number;
  jointRearHandRodY: number;
  jointFrontHandRodX: number;
  jointFrontHandRodY: number;
  shoulderY: number;
  upperArmLen: number;
  foreArmLen: number;
  torsoRise: number;
  torsoHeight: number;
  hipY: number;
  thighLen: number;
  footY: number;
  neckOffsetY: number;
  headTorsoLagFactor: number;
  legVisualScaleMin: number;
};

export type PerformancePuppetTransform = {
  x: number;
  y: number;
  scale: number;
  rotationDeg: number;
};

export const DEFAULT_PERFORMANCE_BONES: PerformanceBoneMetrics = {
  skinScale: 1,
  skinScaleHead: 1,
  skinScaleTorso: 1,
  skinScalePelvis: 1,
  skinScaleLeftUpperArm: 1,
  skinScaleLeftForearm: 1,
  skinScaleRightUpperArm: 1,
  skinScaleRightForearm: 1,
  skinScaleLeftLeg: 1,
  skinScaleRightLeg: 1,
  offsetHeadX: 0,
  offsetHeadY: 0,
  offsetTorsoX: 0,
  offsetTorsoY: 0,
  offsetPelvisX: 0,
  offsetPelvisY: 0,
  offsetLeftUpperArmX: 0,
  offsetLeftUpperArmY: 0,
  offsetLeftForearmX: 0,
  offsetLeftForearmY: 0,
  offsetRightUpperArmX: 0,
  offsetRightUpperArmY: 0,
  offsetRightForearmX: 0,
  offsetRightForearmY: 0,
  offsetLeftLegX: 0,
  offsetLeftLegY: 0,
  offsetRightLegX: 0,
  offsetRightLegY: 0,
  jointShoulderX: 0,
  jointShoulderY: 0,
  jointElbowX: 0,
  jointElbowY: 0,
  jointHipX: 0,
  jointHipY: 0,
  jointKneeX: 0,
  jointKneeY: 0,
  jointLeftShoulderX: 0,
  jointLeftShoulderY: 0,
  jointRightShoulderX: 0,
  jointRightShoulderY: 0,
  jointLeftElbowX: 0,
  jointLeftElbowY: 0,
  jointRightElbowX: 0,
  jointRightElbowY: 0,
  jointLeftHipX: 0,
  jointLeftHipY: 0,
  jointRightHipX: 0,
  jointRightHipY: 0,
  jointLeftKneeX: 0,
  jointLeftKneeY: 0,
  jointRightKneeX: 0,
  jointRightKneeY: 0,
  jointNeckRodX: 0,
  jointNeckRodY: 0,
  jointRearHandRodX: 0,
  jointRearHandRodY: 0,
  jointFrontHandRodX: 0,
  jointFrontHandRodY: 0,
  shoulderY: 80,
  upperArmLen: 65,
  foreArmLen: 55,
  torsoRise: 90,
  torsoHeight: 80,
  hipY: 80,
  thighLen: 65,
  footY: 60,
  neckOffsetY: 10,
  headTorsoLagFactor: 1,
  legVisualScaleMin: 0.75,
};

export const DEFAULT_PERFORMANCE_PUPPET_TRANSFORM: PerformancePuppetTransform = {
  x: 720,
  y: 410,
  scale: 1,
  rotationDeg: 0,
};