// shadow story/src/types/performanceControlPoints.ts
export type PerformanceControlPoints = {
  neckX: number;
  neckY: number;
  rearHandX: number;
  rearHandY: number;
  frontHandX: number;
  frontHandY: number;
  leftShoulderX: number;
  leftShoulderY: number;
  leftElbowX: number;
  leftElbowY: number;
  rightShoulderX: number;
  rightShoulderY: number;
  rightElbowX: number;
  rightElbowY: number;
  leftHipX: number;
  leftHipY: number;
  rightHipX: number;
  rightHipY: number;
  pelvisX: number;
  pelvisY: number;
};

export const DEFAULT_PERFORMANCE_CONTROL_POINTS: PerformanceControlPoints = {
  neckX: 0,
  neckY: 0,
  rearHandX: 0,
  rearHandY: 0,
  frontHandX: 0,
  frontHandY: 0,
  leftShoulderX: 0,
  leftShoulderY: 0,
  leftElbowX: 0,
  leftElbowY: 0,
  rightShoulderX: 0,
  rightShoulderY: 0,
  rightElbowX: 0,
  rightElbowY: 0,
  leftHipX: 0,
  leftHipY: 0,
  rightHipX: 0,
  rightHipY: 0,
  pelvisX: 0,
  pelvisY: 0,
};